EIB-020: Default settings of Electronic In Box. Part 3. Phase 2
