+++ src/main/java/com/ctk/api/ElectronicInBoxLoginRedirect.java
@@ -1,23 +1,23 @@
 package com.ctk.api;
 
-import com.ctk.dao.GrayScaleDao;
+import com.ctk.dao.GrayScale;
 
 import javax.inject.Inject;
 
 public class ElectronicInBoxLoginRedirect {
 
     @Inject
-    private GrayScaleDao grayScaleDao;
+    private GrayScale grayScale;
 
     public String ElectronicInBoxLoghinRedirect() {
         String electronicInBoxLoginRedirect;
 
         String grayScale_ = "000";
 
-        grayScaleDao.loadGrayScaleFile();
+        grayScale.loadGrayScaleFile();
 
-        if (!(grayScaleDao.getGrayScale() == null)) {
-            grayScale_ = grayScaleDao.getGrayScale();
+        if (!(grayScale.getGrayScale() == null)) {
+            grayScale_ = grayScale.getGrayScale();
         }
 
         electronicInBoxLoginRedirect =
