+++ src/main/java/com/ctk/dao/GrayScale.java
@@ -1,5 +1,6 @@
 package com.ctk.dao;
 
+import javax.enterprise.context.SessionScoped;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
@@ -10,7 +11,8 @@ import java.util.Optional;
 
 import static java.lang.String.*;
 
-public class GrayScaleDao extends com.ctk.model.GrayScale implements Serializable {
+@SessionScoped
+public class GrayScale extends com.ctk.model.GrayScale implements Serializable {
 
     public void loadGrayScaleFile() {
         String line;
