+++ src/main/java/com/ctk/services/Settings.java
@@ -1,0 +1,13 @@
+package com.ctk.services;
+
+import java.nio.file.Path;
+
+public interface Settings {
+    Path getPathLESPcsv();
+
+    Path getPathAdmin();
+
+    Path getPathDatabaseInfo();
+
+    Path getPathGrayScaleInfo();
+}
